requirements.txt
python-dotenv==0.19.2
nextcord==2.0.0a4
psutil==5.8.0
hurry.filesize==0.9
aiohttp==3.7.4
pandas==1.3.5
matplotlib==3.5.1
redditeasy==3.6.3
currency-symbols==2.0.1
ujson==5.0.0
uptime==3.0.1
pytz==2021.3
requests==2.26.0
aiosqlite==0.17.0
emojifier==1.1.1
DiscordUtils==1.3.4
DiscordUtils[voice]
dislash.py==1.4.9
googletrans==4.0.0rc1
ocrspace==2.3.0

